requirements.txt
numpy==1.21.6
pandas==1.3.5
Django==2.1.7
PyMySQL==0.9.3
torch==1.13.1
torchvision==0.14.1
transformers==4.30.2
faiss-cpu==1.7.4
safetensors==0.3.3.post1
datasets==2.13.2
sentence-transformers==2.2.2
sentencepiece==0.1.96




